requirements.txt
streamlit>=1.28.0
opencv-python>=4.8.0
numpy>=1.24.0
scipy>=1.11.0
scikit-image>=0.21.0
Pillow>=10.0.0
plotly>=5.17.0
anthropic>=0.18.0

